Support optional PXF_MASTER_HOSTNAME env var
This env var would override the hostname that the operating system
reports, for users who have, for example, an internal network/hostname
for GPDB and external network/hostname for other things. If the hostname
that the system reports is not known to the GPDB, the user can avoid
errors by setting `PXF_MASTER_HOSTNAME` to their GPDB-facing hostname.