Run dataproc create/destroy locally
Using piper, we can run the Concourse tasks locally. This requires
making some changes to the scripts so that options can be turned on/off.

We want dataproc instances for local use to have

1) an external IP
2) local public ssh key for easy access
3) a petname we provide
4) local user as a hadoop proxy
5) palo alto internal network tag
6) default, not dynamic subnet

Besides all the regular stuff we get in dataproc_env_files, we also want
a snippet we can add to /etc/hosts, which is in
dataproc_env_files/dataproc_hosts.

Authored-by: Oliver Albertini <oalbertini@pivotal.io>