Add test for AvroFileAccessor
We are testing that the AvroFileAccessor#initialize() does the right
thing whether the user provides the path to a DATA-SCHEMA, and whether
or not we are reading or writing to external data source.

Also involves a cleanup of AvroResolver#openExternalSchema(), which is
not really used since the accessor is initialized first.

Authored-by: Oliver Albertini <oalbertini@pivotal.io>