Integration test Avro write feature
Tests create a writable table in GPDB that can export to HCFS in Avro
format. Then gets the Avro files off of HCFS and makes JSON records out
of them. Compares with JSON records that are known to be correct, then
uses a readable table to make sure that the data comes back in via GPDB.

Test cases covered include complex/primitive data types, whether the
user provides a schema file (avro/avsc format) and in different
locations (HCFS or classpath). If the user doesn't provide a schema, we
generate one based on the GPDB table. We also check that the table can
handle nulls when auto-generating schemas (this means that we are using
UNION type everywhere in Avro).

Also backfill some tests in Avro read automation suite. Read should also
be able to handle reading user-provided schemas.

Authored-by: Oliver Albertini <oalbertini@pivotal.io>