Bump greenplum-db/gp-common-go-libs to latest
This version of the libs contains changes that simplify running
things on master standby and executing commands from local (rsync,
useful in pxf cluster sync command).

Also makes it easier to run things on hosts (passing a function that
takes a string instead of contentID to gp-common-go-libs).

Also improves usage of cluster API so we can stop counting hosts in our
code.

Authored-by: Oliver Albertini <oalbertini@pivotal.io>