Address CVE-2020-1938
The AJP protocol is enabled by default in Tomcat versions < 7.0.100,
with the AJP connector listening in TCP port 8009 by default. This is
low impact on PXF because PXF should only be accessible from segments
collocated with the PXF Server. It is important to note that mitigation
is only required if an AJP port is accessible to untrusted users.

Tomcat 7.0.100 disables AJP protocol by default