URL decode custom header values when the X-GP-ENCODED-HEADER-VALUES is true
Starting in Tomcat version 7.0.100, HTTP header validation was tightened
disallowing invalid header values. These changed in Tomcat were a side
effect of fixes for CVE-2020-1935. This issue manifests when a table is
created with non printable delimiters, for example:

    CREATE EXTERNAL TABLE foo ()
    LOCATION ('pxf://some/path?PROFILE=demo&SERVER=demo')
    FORMAT 'CSV' ( DELIMITER E'\x01' );

The delimiter value is passed to PXF Server in an HTTP header, however
unencoded. This is causing Tomcat to drop headers with invalid header
values, and causing PXF to behave incorrectly.

In this commit, we decode custom (X-GP-*) header values when the
X-GP-ENCODED-HEADER-VALUES is true.